refactor: collect and export components from a single index file
