+++ src/components/index.ts
@@ -1,0 +1,35 @@
+import AvatarField from './AvatarField/AvatarField.tsx';
+import Button from './Button/Button.tsx';
+import Checkbox from './Checkbox/Checkbox.tsx';
+import ErrorBoundary from './ErrorBoundary/ErrorBoundary.tsx';
+import ErrorMessage from './ErrorComponent/ErrorMessage.tsx';
+import Fallback from './Fallback/Fallback.tsx';
+import Loader from './Loader/Loader.tsx';
+import MainLayout from './MainLayout/MainLayout.tsx';
+import NoContent from './NoContent/NoContent.tsx';
+import NotFound from './NotFound/NotFound.tsx';
+import ProfileCard from './ProfileCard/ProfileCard.tsx';
+import RadioField from './RadioField/RadioField.tsx';
+import RadioGroup from './RadioGroup/RadioGroup.tsx';
+import RootLayout from './RootLayout/RootLayout.tsx';
+import SelectGroup from './SelectGroup/SelectGroup.tsx';
+import TextField from './TextField/TextField.tsx';
+
+export {
+  AvatarField,
+  Button,
+  Checkbox,
+  ErrorBoundary,
+  ErrorMessage,
+  Fallback,
+  Loader,
+  MainLayout,
+  NoContent,
+  NotFound,
+  ProfileCard,
+  RadioField,
+  RadioGroup,
+  RootLayout,
+  SelectGroup,
+  TextField,
+};
