+++ src/main.tsx
@@ -4,8 +4,7 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router';
 import { createGlobalStyle } from 'styled-components';
 import App from './App.tsx';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
-import Fallback from './components/Fallback/Fallback.tsx';
+import { ErrorBoundary, Fallback } from './components';
 import { setupStore } from './store/store.ts';
 import './index.css';
